added event-deatial end poit and refator update and delete events; added event update and delete utility

+++ src/main/java/com/example/clubsapi/controller/EventController.java
@@ -1,12 +1,17 @@
 package com.example.clubsapi.controller;
 
 
+import com.example.clubsapi.dto.EventDetailDto;
 import com.example.clubsapi.dto.EventDto;
 import com.example.clubsapi.dto.EventResponseDto;
 import com.example.clubsapi.dto.EventUserResponseDto;
+import com.example.clubsapi.entity.UserEntity;
+import com.example.clubsapi.exception.ResousrceNotFoundException;
+import com.example.clubsapi.repository.UserRepository;
 import com.example.clubsapi.services.impl.EventServiceImpl;
+import com.example.clubsapi.utiltiy.AccessControl;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.access.annotation.Secured;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
 
@@ -19,9 +24,13 @@ import java.util.Map;
 public class EventController {
 
     private EventServiceImpl eventService;
+    private UserRepository userRepository;
+    private AccessControl accessControl;
 
-    public EventController(EventServiceImpl eventService) {
+    public EventController(EventServiceImpl eventService, UserRepository userRepository, AccessControl accessControl) {
         this.eventService = eventService;
+        this.userRepository = userRepository;
+        this.accessControl = accessControl;
     }
 
     @PreAuthorize("hasRole('MODERATOR')")
@@ -35,7 +44,7 @@ public class EventController {
         return ResponseEntity.ok(res);
     }
 
-    @PreAuthorize("hasRole('USER')")
+    @PreAuthorize("hasRole('USER') or hasRole('MODERATOR')")
     @PostMapping("/join/{eventName}")
     public ResponseEntity<?> joinEvent(@PathVariable("eventName")String eventname){
         eventService.joinEvent(eventname);
@@ -44,11 +53,20 @@ public class EventController {
         return ResponseEntity.ok(res);
     }
 
-    @PreAuthorize("hasRole('USER') or hasRole('MODERATOR') ")
     @GetMapping("/allevents")
-    public ResponseEntity<?> getEvents(){
+    public ResponseEntity<?> getallEvents(){
         List<EventResponseDto> eventResponse = eventService.getAllEvents();
         return ResponseEntity.ok(eventResponse);
+    }
+
+    @GetMapping("/event-detail/{eventId}")
+    public ResponseEntity<?> getevent(@PathVariable("eventId") int eventId){
+        try{
+            EventDetailDto eventDetailDto =eventService.getEvent(eventId);
+            return ResponseEntity.ok(eventDetailDto);
+        }catch (ResousrceNotFoundException re){
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(re.getMessage());
+        }
     }
 
     @PreAuthorize("hasRole('MODERATOR')")
@@ -61,14 +79,26 @@ public class EventController {
     @PreAuthorize("hasRole('MODERATOR')")
     @PatchMapping("/update/{eventId}")
     public ResponseEntity<?> updateEvent(@PathVariable("eventId")int eventId,@RequestBody EventDto eventDto){
-        EventDto dto = eventService.updateEvent(eventId,eventDto);
-        return ResponseEntity.ok(dto);
+        if(accessControl.updateEventUtility(eventId)){
+            EventDto dto = eventService.updateEvent(eventId,eventDto);
+            return ResponseEntity.ok(dto);
+        }
+        else {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(new ResousrceNotFoundException("403 Error"));
+        }
+
 
     }
     @PreAuthorize("hasRole('MODERATOR')")
     @DeleteMapping("/delete/{eventId}")
     public ResponseEntity<?> deleteEvent(@PathVariable("eventId")int id){
-        eventService.deleteEvent(id);
-        return ResponseEntity.ok().build();
+        if(accessControl.deleteEventUtility(id)){
+            eventService.deleteEvent(id);
+            return ResponseEntity.ok().build();
+        }
+        else {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(new ResousrceNotFoundException("Error 401"));
+        }
+
     }
 }
